\name{wgeesel-internal}
\alias{mat.prod}
\alias{E_i_est}
\alias{newton_raphson}
\alias{us_matrix}
\alias{V_w_est}
\alias{binary_fun}
\alias{cluster.size}
\alias{cluster_est}
\alias{cont_fun}
\alias{cormax.ar1}
\alias{cormax.exch}
\alias{cormax.ind}
\alias{count_fun}
\alias{cs_con_str_est}
\alias{phi_con_str_est}
\title{Internal wgeesel Functions}
\description{
  Internal \pkg{wgeesel} functions.
}
\keyword{internal}